League Picks AJAX call update
Updated to code to no longer use setInterval() with an AJAX call (as
this could have unintended errors if the user's connection is slow).
Instead the "complete" parameter of the AJAX call is used to recursively
call userPicks() and updateNFLScores(). This guarantees that the call
stack will only have 1 AJAX at a time before another one is executed.
The timeout for each call is set to 10 seconds.