League Picks points and wins % fix and code optimization
Fixing a bug that was created in last update when tables on league picks
pages were merge. The table had an id that was used in a jQuery selector
that was necessary for summing the points and calculating the win %.

This ID is now placed on the <tbody> HTML tag and renamed to match the
old table's ID.

Updated uncessary deleting of the NFL scores and quarters in order to
update the scores. Instead a jQuery selector is used to find all
home/away scores and quarters and the text is the only thing updated.
This way we only have to worry about the formatting once (upon loading),
and after that the text in each cell is simply updated.