WeekGames object update
Updated code to use WeekGames object. This object simplifies a lot of
the code as it contains all properties needed to determine user's
scores, to populate user's pick's, and to update NFL scores on League
Picks page.

Also updated code to catch error when attempting to get timenow from api
service. When it fails, a timeout is set to rerun the code after 5
seconds. This will "hopefully" eliminate the bug where all games are
incorrectly marked on current NFL week.

Various code clean up and fixes.