Fix: Modal popup for non-current weeks
Fixed an issued introduced in last update, where when a game is clicked
on "league picks" page, it will update "weekData", so that picks can
reference most recent information. "weekData" should only be updated
when the current week is equal to the selected webpage. An "If-Else"
statement is implemented for this fix.