Fix: Overlap issues on User Picks page
Fixed an overlapping issue on the user picks page with the table div,
week choice dropdown, and show bye teams button. The issue came after
change the table's parent div's style to "position: relative". Because
of this a margin of 40px needed to be added to prevent the overlapping
issue causing the button/dropdown to be unclickable.