Fix: Red background disappear if no pick made
Fixed a bug in updateUserPicks function where the background would be
removed if the user made no picks. Also, updated the function to skip
cells that already have a background (as this means the pick was marked
correct/incorrect previously).

Changed "weekData" to be global, so that showing user picks on mobile
devices would load faster (no having to wait on an AJAX call to
NFL.com). This all cuts down on the number of AJAX calls.